+++ src/lw_common/util/simple_cache_dictionary.cs
@@ -45,5 +45,11 @@ namespace lw_common {
             }
         }
 
+        public void clear() {
+            lock (this) {
+                values_.Clear();
+            }
+        }
+
     }
 }